+++ .pre-commit-config.yaml
@@ -1,43 +1,43 @@
-repos: 
--  repo: https://github.com/timothycrosley/isort
-   rev: 5.2.0
-   hooks:
-   -   id: sort
-       types: [python]
-       args: ["--settings-path", "setup.cfg"]
--  repo: https://github.com/psf/black
-   rev: 19.10b0
-   hooks: 
-   -   id: black
-       language_version: python3.7
--  repo: local
-   hooks: 
-   -   id: flake8
-       name: flake8
-       entry: flake8
-       args: ["--config=setup.cfg"]
-       additional_dependencies: [flake8-bandit, flake8-bugbear, flake8-builtins, flake8-comprehensions]
-       language: python
-       types: [python]
--  repo: https://github.com/pre-commit/pre-commit-hooks
-   rev: v3.1.0
-   hooks: 
-   - id: check-added-large-files
-   - id: check-merge-conflict
-   - id: check-yaml
-   - id: debug-statements
-   - id: detect-private-key
-   - id: end-of-file-fixer
-   - id: requirements-txt-fixer
-   - id: trailing-whitespace
- - repo: https://github.com/pre-commit/pygrep-hooks
-   rev: v1.5.1
-   hooks:
-   - id: python-check-mock-methods
-   - id: python-use-type-annotations
-   - id: python-no-eval
-- repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.782
-  hooks:
-  - id: mypy
-    args: ["--config-file", "setup.cfg"]
+repos:
+  - repo: https://github.com/pycqa/isort
+    rev: 5.5.2
+    hooks:
+      - id: isort
+        types: [python]
+        args: ["--settings-path", "setup.cfg"]
+  - repo: https://github.com/psf/black
+    rev: 19.10b0
+    hooks:
+      - id: black
+        language_version: python3.6
+  - repo: local
+    hooks:
+      - id: flake8
+        name: flake8
+        entry: flake8
+        args: ["--config=setup.cfg"]
+        additional_dependencies: [flake8-bandit, flake8-bugbear, flake8-builtins, flake8-comprehensions]
+        language: python
+        types: [python]
+  - repo: https://github.com/pre-commit/pre-commit-hooks  
+    rev: v3.1.0
+    hooks:
+      - id: check-added-large-files 
+      - id: check-merge-conflict
+      - id: check-yaml
+      - id: debug-statements
+      - id: detect-private-key
+      - id: end-of-file-fixer
+      - id: requirements-txt-fixer
+      - id: trailing-whitespace
+  - repo: https://github.com/pre-commit/pygrep-hooks    
+    rev: v1.5.1
+    hooks:
+      - id: python-check-mock-methods
+      - id: python-use-type-annotations
+      - id: python-no-eval
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.782
+    hooks:
+      - id: mypy
+        args: ["--config-file", "setup.cfg"]
