+++ setup.cfg
@@ -5,7 +5,7 @@ select = A,B, B9, C, E, F, S, W
 ignore = E203, E501, E722, W503
 per-file-ignore= **/test*.py:S101
 
-[isort] 
+[isort]
 profile=black
 
 [mypy]
